added sample test data
Imagine a grid of lines: At each intersection of lines, there may or may not be a node. For a grid of any size, with any number of fixed nodes on it, design the algorithm which will output the location as a coordinate for each node, and the distance to its nearest neighbour to the right, (if one exists) and its nearest neighbour below, if one exists. Please write this as a program in Java, in such a way that other people can understand it, it can easily be maintained or extended, and it is as efficient as possible. Please also ensure that it is easily testable (ideally via unit tests) and that it allows any set of nodes and grid size to be set-up